requirements.txt
Pillow
filetype
matplotlib
scipy
soundfile